@@ -1,51 +1,55 @@
 {
-  "name": "frontend",
-  "version": "0.0.0",
-  "license": "MIT",
-  "scripts": {
-    "start": "ng serve --proxy-config proxy.conf.json",
-    "build": "ng build --dev",
-    "prod": "ng build --prod",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "ngx-bootstrap": "^2.0.0-rc.0",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
-  },
-  "devDependencies": {
-    "@angular/cli": "1.6.1",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
-  }
+    "name": "WebDB",
+    "version": "4.0.0",
+    "license": "MIT",
+    "scripts": {
+        "start": "ng serve --proxy-config proxy.conf.json",
+        "build": "ng build --dev",
+        "prod": "ng build --prod",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^5.2.0",
+        "@angular/common": "^5.2.0",
+        "@angular/compiler": "^5.2.0",
+        "@angular/core": "^5.2.0",
+        "@angular/forms": "^5.2.0",
+        "@angular/http": "^5.2.0",
+        "@angular/platform-browser": "^5.2.0",
+        "@angular/platform-browser-dynamic": "^5.2.0",
+        "@angular/router": "^5.2.0",
+        "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.9",
+        "bootstrap": "^4.0.0-beta.2",
+        "caniuse-db": "^1.0.30000791",
+        "chart.js": "^2.7.1",
+        "core-js": "^2.5.3",
+        "font-awesome": "^4.7.0",
+        "ng2-charts": "^1.6.0",
+        "rxjs": "^5.5.6",
+        "zone.js": "^0.8.20"
+    },
+    "devDependencies": {
+        "@angular/cli": "^1.6.4",
+        "@angular/compiler-cli": "^5.2.0",
+        "@angular/language-service": "^5.2.0",
+        "@types/jasmine": "~2.6.3",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/node": "^8.5.8",
+        "codelyzer": "^4.0.2",
+        "jasmine-core": "~2.8.0",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "~1.7.0",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-cli": "~1.0.1",
+        "karma-coverage-istanbul-reporter": "^1.3.3",
+        "karma-jasmine": "^1.1.1",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.2.0",
+        "ts-node": "~3.3.0",
+        "tslint": "~5.8.0",
+        "typescript": "~2.4.2"
+    }
 }
