@@ -1,15 +1,14 @@
-import {Component, OnInit} from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss']
 })
-export class AppComponent implements OnInit
-{
-  constructor() {}
+export class AppComponent implements OnInit {
+    constructor() {
+    }
 
-  ngOnInit(): void
-  {
-  }
+    ngOnInit() {
+    }
 }
